+++ NEWS.md
@@ -5,6 +5,7 @@
 - Add `filterMsLevel` method.
 - Add `filterRt` method.
 - Add `filterDataOrigin` method.
+- Add `filterPrecursorMzRange` method.
 
 ## Changes in 0.0.4